arrodes_unified/requirements.txt
# Discord
discord.py>=2.3.2

# LLM Providers
google-genai>=0.2.0
groq

# RAG
faiss-cpu
rank-bm25
numpy

# API
flask

# Utils
python-dotenv>=1.0.1
requests
